+++ src/System/Taffybar/Widget.hs
@@ -9,6 +9,9 @@ module System.Taffybar.Widget
   -- * "System.Taffybar.Widget.CommandRunner"
   , module System.Taffybar.Widget.CommandRunner
 
+  -- * "System.Taffybar.Widget.Crypto"
+  , module System.Taffybar.Widget.Crypto
+
   -- * "System.Taffybar.Widget.Decorators"
   , module System.Taffybar.Widget.Decorators
 
@@ -64,6 +67,7 @@ module System.Taffybar.Widget
 import System.Taffybar.Widget.Battery
 import System.Taffybar.Widget.CPUMonitor
 import System.Taffybar.Widget.CommandRunner
+import System.Taffybar.Widget.Crypto
 import System.Taffybar.Widget.Decorators
 import System.Taffybar.Widget.DiskIOMonitor
 import System.Taffybar.Widget.FSMonitor